Tablas

+++ POS/src/POS.Application/Interfaces/IUserEcoApplication.cs
@@ -7,6 +7,6 @@ namespace POS.Application.Interfaces;
 public interface IUserEcoApplication
 {
     Task<BaseResponse<IEnumerable<UserEco>>> ListSelectUser();
-    Task<BaseResponse<UserEco>> GetUserById(int id);
+    Task<BaseResponse<UserEco>> GetUserById(short id);
     Task<BaseResponse<short>> RegisterUser(UserEcoRequestDto addUser);
 }
